+++ packages/calcite-components/src/components/shell/shell.tsx
@@ -18,6 +18,7 @@ import { CSS, SLOTS } from "./resources";
  * @slot center-row - [Deprecated] Use the `"panel-bottom"` slot instead. A slot for adding the bottom `calcite-shell-center-row`.
  * @slot modals - A slot for adding `calcite-modal` components. When placed in this slot, the modal position will be constrained to the extent of the shell.
  * @slot alerts - A slot for adding `calcite-alert` components. When placed in this slot, the alert position will be constrained to the extent of the shell.
+ * @slot sheets - A slot for adding `calcite-sheet` components. When placed in this slot, the sheet position will be constrained to the extent of the shell.
  */
 
 @Component({
@@ -53,6 +54,8 @@ export class Shell implements ConditionalSlotComponent {
 
   @State() hasModals = false;
 
+  @State() hasSheets = false;
+
   // --------------------------------------------------------------------------
   //
   //  Lifecycle
@@ -86,6 +89,15 @@ export class Shell implements ConditionalSlotComponent {
     slotChangeGetAssignedElements(event)?.map((el) => {
       if (el.nodeName === "CALCITE-ALERT") {
         (el as HTMLCalciteAlertElement).slottedInShell = true;
+      }
+    });
+  };
+
+  handleSheetsSlotChange = (event: Event): void => {
+    this.hasSheets = !!slotChangeHasAssignedElement(event);
+    slotChangeGetAssignedElements(event)?.map((el) => {
+      if (el.nodeName === "CALCITE-SHEET") {
+        (el as HTMLCalciteSheetElement).slottedInShell = true;
       }
     });
   };
@@ -125,6 +137,14 @@ export class Shell implements ConditionalSlotComponent {
     return (
       <div hidden={!this.hasAlerts}>
         <slot key="alerts" name={SLOTS.alerts} onSlotchange={this.handleAlertsSlotChange} />
+      </div>
+    );
+  }
+
+  renderSheets(): VNode {
+    return (
+      <div hidden={!this.hasSheets}>
+        <slot key="sheets" name={SLOTS.sheets} onSlotchange={this.handleSheetsSlotChange} />
       </div>
     );
   }
@@ -191,6 +211,7 @@ export class Shell implements ConditionalSlotComponent {
       <div class={CSS.positionedSlotWrapper}>
         {this.renderAlerts()}
         {this.renderModals()}
+        {this.renderSheets()}
       </div>
     );
   }