+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.0-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.12...@esri/calcite-components-react@1.7.0-next.13) (2023-08-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.11...@esri/calcite-components-react@1.7.0-next.12) (2023-08-29)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.10...@esri/calcite-components-react@1.7.0-next.11) (2023-08-28)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.10](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.9...@esri/calcite-components-react@1.7.0-next.10) (2023-08-28)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.9](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.8...@esri/calcite-components-react@1.7.0-next.9) (2023-08-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.8](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.7...@esri/calcite-components-react@1.7.0-next.8) (2023-08-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.7](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.6...@esri/calcite-components-react@1.7.0-next.7) (2023-08-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.6](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.5...@esri/calcite-components-react@1.7.0-next.6) (2023-08-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.5](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.4...@esri/calcite-components-react@1.7.0-next.5) (2023-08-24)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [1.7.0-next.4](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.3...@esri/calcite-components-react@1.7.0-next.4) (2023-08-23)
 
 **Note:** Version bump only for package @esri/calcite-components-react